+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [0.3.2]
 * `PosColumn` can contain encoded text (`textEncoded` field)
+* Bug fix: Columns with `PosTextSize` > `size1`
+* Added Barcode Code128
+* Added new code pages
+* `imageRaster` bug fixed
+* Ticket bytecode optimization: do not generate align left command (it's a default value)
+* Added new image print function: `GS ( L`
 
 
 ## [0.3.1]